# docs/admin/swagger.yaml
basePath: /api/v1
definitions:
  admin_accountuser.ReqCreateUser:
    properties:
      email:
        description: 邮箱
        example: abc@a.com
        type: string
      is_admin:
        description: 是否属于管理员账号
        example: true
        type: boolean
      nickname:
        description: 昵称
        example: Tom
        type: string
      password:
        description: 密码
        example: "123456"
        type: string
      password_confirm:
        description: 确认密码
        example: "123456"
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
    required:
    - password
    - password_confirm
    - phone
    type: object
  admin_accountuser.ReqEditUser:
    properties:
      email:
        description: 邮箱
        example: abc@a.com
        type: string
      is_admin:
        description: 是否属于管理员账号
        example: true
        type: boolean
      nickname:
        description: 昵称
        example: Tom
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
      status:
        description: 用户状态
        example: FREEZE
        type: string
      uuid:
        type: string
    required:
    - phone
    - uuid
    type: object
  admin_accountuser.ReqUpdateUser:
    properties:
      avatar:
        description: 用户头像URL
        example: a/b/c.jpg
        type: string
      email:
        description: 邮箱
        example: abc@abc.com
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/model.UserGender'
        description: 性别:3-保密,1-男,2-女
        example: 3
      nickname:
        description: 昵称
        example: Tom
        type: string
      signature:
        description: 个性化签名
        example: haha
        type: string
      status:
        description: 用户状态
        example: FREEZE
        type: string
    type: object
  admin_accountuser.ReqUserUUID:
    properties:
      uuid:
        type: string
    required:
    - uuid
    type: object
  admin_accountuser.ReqUserUUIDs:
    properties:
      uuids:
        items:
          type: string
        type: array
    required:
    - uuids
    type: object
  admin_accountuser.RespUserDetail:
    properties:
      create_time:
        description: 账号创建时间(unix秒时间戳)
        example: 1724914598
        type: integer
      email:
        description: 邮箱
        example: abc@abc.com
        type: string
      is_admin:
        description: 是否管理员
        example: false
        type: boolean
      last_login_time:
        description: 最近登录时间(unix秒时间戳)
        example: 1724914598
        type: integer
      nickname:
        description: 昵称
        example: Tom
        type: string
      phone:
        description: 手机号
        example: "13800138000"
        type: string
      status:
        allOf:
        - $ref: '#/definitions/model.UserStatusType'
        description: 用户状态
        example: ACTIVE
      uuid:
        description: 用户UUID,32位字符串
        example: 826d6b1aa64d471d822d667e92218158
        type: string
    type: object
...
  goal-app_pkg_render.JsonDataResp:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  goal-app_pkg_render.RespPageJson:
    properties:
      limit:
        type: integer
      page:
        type: integer
      result: {}
      total:
        type: integer
    type: object
  model.UserGender:
    enum:
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - UserGenderMale
    - UserGenderFemale
    - UserGenderUnknown
  model.UserStatusType:
    enum:
    - INACTIVE
    - ACTIVE
    - FREEZE
    - DELETE
    type: string
    x-enum-varnames:
    - UserStatusInactive
    - UserStatusActive
    - UserStatusFreeze
    - UserStatusDelete
info:
  contact: {}
  description: http状态码是200，code为0时表示正常返回；code不为0时表示有业务错误。
  title: 后台管理系统接口文档
  version: "1.0"
paths:
...
  /account/user/{uuid}/profile:
    get:
      description: 用户个人资料
      parameters:
      - description: 用户UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
      security:
      - AdminAuth: []
      summary: 用户个人资料
      tags:
      - 用户管理
  /account/user/create:
    post:
      consumes:
      - application/json
      description: 创建新用户
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin_accountuser.ReqCreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: code不为0时表示错误
          schema:
            allOf:
            - $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
            - properties:
                data:
                  $ref: '#/definitions/admin_accountuser.RespUserDetail'
              type: object
        "500":
          description: Internal Server Error
      security:
      - AdminAuth: []
      summary: 创建用户
      tags:
      - 用户管理
  /account/user/delete:
    post:
      consumes:
      - application/json
      description: 删除单个用户
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin_accountuser.ReqUserUUIDs'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
      security:
      - AdminAuth: []
      summary: 删除用户
      tags:
      - 用户管理
  /account/user/detail:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取用户详情
      parameters:
      - description: 用户UUID
        in: query
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/admin_accountuser.RespUserDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
      security:
      - AdminAuth: []
      summary: 通过用户UUID获取用户详情
      tags:
      - 用户管理
  /account/user/edit:
    post:
      consumes:
      - application/json
      description: 修改用户基本信息
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin_accountuser.ReqEditUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
      security:
      - AdminAuth: []
      summary: 修改用户基本信息
      tags:
      - 用户管理
  /account/user/list:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取用户列表
      parameters:
      - description: 数据创建结束区间
        in: query
        name: create_time_end
        type: integer
      - description: 数据创建开始区间
        in: query
        name: create_time_start
        type: integer
      - description: 邮箱,模糊查询
        example: abc@abc.com
        in: query
        name: email
        type: string
      - description: 是否admin, 1 or 0
        example: 1
        in: query
        name: is_admin
        type: integer
      - description: 最近登录结束时间
        in: query
        name: last_login_time_end
        type: integer
      - description: 最近登录开始时间
        in: query
        name: last_login_time_start
        type: integer
      - default: 10
        description: 每页数量
        example: 10
        in: query
        name: limit
        type: integer
      - description: 昵称,模糊查询
        example: Tom
        in: query
        name: nickname
        type: string
      - default: 1
        description: 页码
        example: 1
        in: query
        name: page
        type: integer
      - description: 手机号,模糊查询
        example: "13800138000"
        in: query
        name: phone
        type: string
      - collectionFormat: multi
        description: 用户状态
        example:
        - FREEZE
        - ACTIVE
        in: query
        items:
          enum:
          - INACTIVE
          - ACTIVE
          - FREEZE
          - DELETE
          type: string
        name: status[]
        type: array
      - description: 用户UUID,精确匹配
        example: 826d6b1aa64d471d822d667e92218158
        in: query
        name: uuid
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code不为0时表示有错误
          schema:
            allOf:
            - $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/goal-app_pkg_render.RespPageJson'
                  - properties:
                      result:
                        items:
                          $ref: '#/definitions/admin_accountuser.RespUserDetail'
                        type: array
                    type: object
              type: object
        "500":
          description: Internal Server Error
      security:
      - AdminAuth: []
      summary: 获取用户列表
      tags:
      - 用户管理
  /account/user/reset-password:
    post:
      consumes:
      - application/json
      description: 将用户密码重置为123456
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin_accountuser.ReqUserUUID'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
      security:
      - AdminAuth: []
      summary: 将用户密码重置为123456
      tags:
      - 用户管理
  /account/user/update:
    post:
      consumes:
      - application/json
      description: 更新用户数据
      parameters:
      - description: 请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/admin_accountuser.ReqUpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
      security:
      - AdminAuth: []
      summary: 更新用户
      tags:
      - 用户管理
  /account/users/{uuid}/profile/update:
    post:
      consumes:
      - application/json
      description: 更新用户个人资料
      parameters:
      - description: 用户UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/goal-app_pkg_render.JsonDataResp'
      security:
      - AdminAuth: []
      summary: 删除用户
      tags:
      - 用户管理
...
securityDefinitions:
  AdminAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"